Fix workflow for populating configurable values
This commit fixes the bug that arose when configuring the plugin. The
licence type cannot be selected until the licence types are fetched from
the API, and they can't be fetched from the API until the API key is
provided. This is now fixed so after the API key is provided, the
licence types are fetched and one can be selected.

We also fix errant tabs